backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication and security
bcrypt==4.0.1
PyJWT==2.8.0

# Database
pymongo==4.6.0

# AI/ML
google-generativeai==0.3.2

# PDF processing
pypdf==3.17.1

# Environment variables
python-dotenv==1.0.0

# Email validation (required by Pydantic)
email-validator==2.1.0

# For production deployment (optional)
gunicorn==21.2.0